moved animated messaging on transaction page to its own component to share with balance page.
