+++ frontend/src/layouts/TransactionPage.tsx
@@ -1,13 +1,12 @@
 import { standardFormClasses } from "../styles/styles";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { usdInputMask } from "../utils";
 import { useModal } from "../context/ModalProvider";
-import { motion } from "framer-motion";
+import MessagingHeaders from "../components/MessagingHeader";
 import useAuth from "../hooks/useAuth";
 import PageWrapper from "./PageWrapper";
 import useAccounts from "../hooks/useAccount";
-import AppearingText from "../components/AppearingText";
 import StandardInput from "../components/StandardInput";
 import StandardButton from "../components/StandardButton";
 
@@ -15,7 +14,6 @@ export default function TransactionPage() {
   const navigateTo = useNavigate();
   const { openConfirmModal } = useModal();
   const { currentUser } = useAuth();
-  const messageRef = useRef<string[]>(messaging[Math.floor(Math.random() * 5)]);
 
   useEffect(() => {
     if (!currentUser) navigateTo("/login");
@@ -54,21 +52,7 @@ export default function TransactionPage() {
   return (
     <PageWrapper>
       <div className={standardFormClasses}>
-        <AppearingText
-          words={messageRef.current[0]}
-          className="text-3xl bold -translate-y-4 -translate-x-8"
-        />
-        <motion.div
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          transition={{ delay: 1 }}
-        >
-          <AppearingText
-            words={messageRef.current[1]}
-            className="text-2xl bold -translate-y-4 translate-x-8"
-          />
-        </motion.div>
-
+        <MessagingHeaders />
         <p className="text-lg mt-4">
           {`How much would you like to ${pathname.slice(
             1,
@@ -95,11 +79,3 @@ export default function TransactionPage() {
     </PageWrapper>
   );
 }
-
-const messaging = [
-  ["Need a loan?", "We're here for you."],
-  ["Looking to invest?", "Let us guide you."],
-  ["Trying to save?", "Let's make a plan."],
-  ["Want financial advice?", "You can count on us."],
-  ["Building credit?", "We can support you."],
-];
